refactor : 공통 응답 포맷 변경에 따른 HomeController, PostLikeController 응답 변경
